[IR Runtime] mapping label_id to label_name in runtime sink
