+++ research/query_service/ir/runtime/src/process/operator/sink/mod.rs
@@ -1,0 +1,40 @@
+//
+//! Copyright 2021 Alibaba Group Holding Limited.
+//!
+//! Licensed under the Apache License, Version 2.0 (the "License");
+//! you may not use this file except in compliance with the License.
+//! You may obtain a copy of the License at
+//!
+//! http://www.apache.org/licenses/LICENSE-2.0
+//!
+//! Unless required by applicable law or agreed to in writing, software
+//! distributed under the License is distributed on an "AS IS" BASIS,
+//! WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//! See the License for the specific language governing permissions and
+//! limitations under the License.
+mod sink;
+
+use ir_common::error::ParsePbError;
+use ir_common::generated::algebra as algebra_pb;
+use ir_common::generated::results as result_pb;
+use pegasus::api::function::MapFunction;
+
+use crate::error::FnGenResult;
+use crate::process::record::Record;
+
+pub trait SinkFunctionGen {
+    fn gen_sink(self) -> FnGenResult<Box<dyn MapFunction<Record, result_pb::Results>>>;
+}
+
+impl SinkFunctionGen for algebra_pb::logical_plan::Operator {
+    fn gen_sink(self) -> FnGenResult<Box<dyn MapFunction<Record, result_pb::Results>>> {
+        if let Some(opr) = self.opr {
+            match opr {
+                algebra_pb::logical_plan::operator::Opr::Sink(sink) => sink.gen_sink(),
+                _ => Err(ParsePbError::from("algebra_pb op is not a sink op"))?,
+            }
+        } else {
+            Err(ParsePbError::EmptyFieldError("algebra op is empty".to_string()))?
+        }
+    }
+}
