+++ research/query_service/ir/runtime/src/process/operator/sink/sink.rs
@@ -1,0 +1,194 @@
+//
+//! Copyright 2021 Alibaba Group Holding Limited.
+//!
+//! Licensed under the Apache License, Version 2.0 (the "License");
+//! you may not use this file except in compliance with the License.
+//! You may obtain a copy of the License at
+//!
+//! http://www.apache.org/licenses/LICENSE-2.0
+//!
+//! Unless required by applicable law or agreed to in writing, software
+//! distributed under the License is distributed on an "AS IS" BASIS,
+//! WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//! See the License for the specific language governing permissions and
+//! limitations under the License.
+
+use std::collections::HashMap;
+use std::convert::TryInto;
+use std::ops::Deref;
+
+use ir_common::generated::algebra as algebra_pb;
+use ir_common::generated::algebra::sink::MapType;
+use ir_common::generated::common as common_pb;
+use ir_common::generated::results as result_pb;
+use ir_common::NameOrId;
+use pegasus::api::function::{FnResult, MapFunction};
+
+use crate::error::FnGenResult;
+use crate::graph::element::{Edge, GraphElement, Vertex, VertexOrEdge};
+use crate::process::operator::sink::SinkFunctionGen;
+use crate::process::record::{Entry, ObjectElement, Record, RecordElement};
+
+#[derive(Debug)]
+pub struct RecordSinkEncoder {
+    /// the given column tags to sink;
+    sink_keys: Vec<NameOrId>,
+    /// flag of sink head
+    is_output_head: bool,
+    /// A map from id to name; including type of Entity (Vertex in Graph Database),
+    /// Relation (Edge in Graph Database) and Column (Property in Graph Database)
+    schema_map: Option<HashMap<(MapType, i32), String>>,
+}
+
+// sink head by default
+impl Default for RecordSinkEncoder {
+    fn default() -> Self {
+        RecordSinkEncoder { sink_keys: vec![], is_output_head: true, schema_map: None }
+    }
+}
+
+impl RecordSinkEncoder {
+    fn entry_to_pb(&self, e: &Entry) -> result_pb::Entry {
+        let inner = match e {
+            Entry::Element(element) => {
+                let element_pb = self.element_to_pb(element);
+                Some(result_pb::entry::Inner::Element(element_pb))
+            }
+            Entry::Collection(collection) => {
+                let mut collection_pb = Vec::with_capacity(collection.len());
+                for element in collection.into_iter() {
+                    let element_pb = self.element_to_pb(element);
+                    collection_pb.push(element_pb);
+                }
+                Some(result_pb::entry::Inner::Collection(result_pb::Collection {
+                    collection: collection_pb,
+                }))
+            }
+        };
+        result_pb::Entry { inner }
+    }
+
+    fn element_to_pb(&self, e: &RecordElement) -> result_pb::Element {
+        let inner = match e {
+            RecordElement::OnGraph(vertex_or_edge) => match vertex_or_edge {
+                VertexOrEdge::V(v) => {
+                    let vertex_pb = self.vertex_to_pb(v);
+                    Some(result_pb::element::Inner::Vertex(vertex_pb))
+                }
+                VertexOrEdge::E(e) => {
+                    let edge_pb = self.edge_to_pb(e);
+                    Some(result_pb::element::Inner::Edge(edge_pb))
+                }
+            },
+            RecordElement::OffGraph(o) => match o {
+                ObjectElement::None => None,
+                ObjectElement::Prop(obj) | ObjectElement::Agg(obj) => {
+                    Some(result_pb::element::Inner::Object(obj.clone().into()))
+                }
+                ObjectElement::Count(cnt) => {
+                    let item = if *cnt <= (i64::MAX as u64) {
+                        common_pb::value::Item::I64(*cnt as i64)
+                    } else {
+                        common_pb::value::Item::Blob(cnt.to_be_bytes().to_vec())
+                    };
+                    let value_pb = common_pb::Value { item: Some(item) };
+                    Some(result_pb::element::Inner::Object(value_pb))
+                }
+            },
+        };
+        result_pb::Element { inner }
+    }
+
+    fn label_to_pb(&self, label: NameOrId, t: MapType) -> common_pb::NameOrId {
+        let mapped_label = if let Some(schema_map) = self.schema_map.as_ref() {
+            match label {
+                NameOrId::Str(_) => label,
+                NameOrId::Id(id) => {
+                    let str_label = schema_map.get(&(t, id)).unwrap();
+                    NameOrId::Str(str_label.clone())
+                }
+            }
+        } else {
+            label
+        };
+        mapped_label.into()
+    }
+
+    fn vertex_to_pb(&self, v: &Vertex) -> result_pb::Vertex {
+        result_pb::Vertex {
+            id: v.id() as i64,
+            label: v
+                .label()
+                .map(|label| self.label_to_pb(label.clone(), MapType::Entity)),
+            // TODO: return detached vertex without property for now
+            properties: vec![],
+        }
+    }
+
+    fn edge_to_pb(&self, e: &Edge) -> result_pb::Edge {
+        result_pb::Edge {
+            id: e.id() as i64,
+            label: e
+                .label()
+                .map(|label| self.label_to_pb(label.clone(), MapType::Relation)),
+            src_id: e.src_id as i64,
+            src_label: e
+                .get_src_label()
+                .map(|label| self.label_to_pb(label.clone(), MapType::Entity)),
+            dst_id: e.dst_id as i64,
+            dst_label: e
+                .get_dst_label()
+                .map(|label| self.label_to_pb(label.clone(), MapType::Entity)),
+            // TODO: return detached edge without property for now
+            properties: vec![],
+        }
+    }
+}
+
+impl MapFunction<Record, result_pb::Results> for RecordSinkEncoder {
+    fn exec(&self, input: Record) -> FnResult<result_pb::Results> {
+        let mut sink_columns = Vec::with_capacity(self.sink_keys.len());
+        if self.is_output_head {
+            let entry = input.get(None);
+            let entry_pb = entry.map(|entry| self.entry_to_pb(entry.deref()));
+            let column_pb = result_pb::Column { name_or_id: None, entry: entry_pb };
+            sink_columns.push(column_pb);
+        }
+        for sink_key in self.sink_keys.iter() {
+            let entry = input.get(Some(sink_key));
+            let entry_pb = entry.map(|entry| self.entry_to_pb(entry.deref()));
+            let column_pb = result_pb::Column {
+                name_or_id: Some(common_pb::NameOrId::from(sink_key.clone())),
+                entry: entry_pb,
+            };
+            sink_columns.push(column_pb);
+        }
+
+        let record_pb = result_pb::Record { columns: sink_columns };
+        let results = result_pb::Results { inner: Some(result_pb::results::Inner::Record(record_pb)) };
+        Ok(results)
+    }
+}
+
+impl SinkFunctionGen for algebra_pb::Sink {
+    fn gen_sink(self) -> FnGenResult<Box<dyn MapFunction<Record, result_pb::Results>>> {
+        let sink_keys = self
+            .tags
+            .into_iter()
+            .map(|tag| tag.try_into())
+            .collect::<Result<_, _>>()?;
+        let is_output_head = self.sink_current;
+        let mut schema_map = HashMap::new();
+        for id_name_mapping_pb in self.id_name_mapping {
+            let map_type = unsafe { ::std::mem::transmute(id_name_mapping_pb.map_type) };
+            schema_map.insert((map_type, id_name_mapping_pb.id), id_name_mapping_pb.name);
+        }
+        let record_sinker = RecordSinkEncoder {
+            sink_keys,
+            is_output_head,
+            schema_map: if schema_map.is_empty() { None } else { Some(schema_map) },
+        };
+        debug!("Runtime sink operator: {:?}", record_sinker);
+        Ok(Box::new(record_sinker))
+    }
+}
